Fix file paths after cherry-picking from cosmic-core
