+++ marvin/deployDataCenter.py
@@ -8,13 +8,16 @@ class DeleteDataCenters: Deletes a DataCenter based upon the dc cfg
                          configuration from DeployDataCenters post
                          its success
 """
-from marvin import configGenerator
-from marvin.cloudstackException import InvalidParameterException
-from marvin.cloudstackAPI import *
-from marvin.codes import (FAILED, SUCCESS)
-from marvin.lib.utils import (random_gen)
-from marvin.config.test_data import test_data
-from sys import exit
+import configGenerator
+from marvinLog import MarvinLog
+from cloudstackException import InvalidParameterException
+from codes import (FAILED, SUCCESS)
+from lib.utils import random_gen
+from config.test_data import test_data
+from cloudstackTestClient import CSTestClient
+from cloudstackAPI import *
+
+import sys
 import os
 import pickle
 from time import sleep, strftime, localtime
@@ -29,14 +32,12 @@ class DeployDataCenters(object):
             the DataCenter settings to an obj file
             ( can be used if wanted to delete the created DC)
     '''
-
     def __init__(self, test_client, cfg):
-        self.__testClient = test_client
+        self.__test_client = test_client
         self.__config = cfg
-        self.__tcRunLogger = MarvinLog('marvin').getLogger()
+        self.__logger = MarvinLog('marvin').getLogger()
         self.__apiClient = None
         self.__cleanUp = {}
-
 
     def __persistDcConfig(self):
         try:
@@ -45,29 +46,27 @@ class DeployDataCenters(object):
             file_to_write = open(dc_file_path, 'w')
             if file_to_write:
                 pickle.dump(self.__cleanUp, file_to_write)
-                self.__tcRunLogger.info("=== Data Center Settings are dumped to %s ===" % dc_file_path)
+                self.__logger.info("=== Data Center Settings are dumped to %s ===" % dc_file_path)
         except Exception as e:
-            self.__tcRunLogger.exception("=== Persisting DataCenter config failed: %s ===" % e)
-
+            self.__logger.exception("=== Persisting DataCenter config failed: %s ===" % e)
 
     def __cleanAndExit(self):
         try:
-            self.__tcRunLogger.info("=== deploy dc failed, so cleaning the created entries ===")
+            self.__logger.info("=== deploy dc failed, so cleaning the created entries ===")
             if not test_data.get("deleteDC", None):
-                self.__tcRunLogger.info("=== Deploy DC Clean Up flag not set. So, exiting ===")
+                self.__logger.info("=== Deploy DC Clean Up flag not set. So, exiting ===")
                 exit(1)
-            self.__tcRunLogger.info("=== Deploy DC Failed, So Cleaning to Exit ===")
-            remove_dc_obj = DeleteDataCenters(self.__testClient, dc_cfg=self.__cleanUp)
+            self.__logger.info("=== Deploy DC Failed, So Cleaning to Exit ===")
+            remove_dc_obj = DeleteDataCenters(self.__test_client, self.__cleanUp)
             if remove_dc_obj:
                 if remove_dc_obj.removeDataCenter() == FAILED:
-                    self.__tcRunLogger.error("=== Removing DataCenter Failed ===")
+                    self.__logger.error("=== Removing DataCenter Failed ===")
                 else:
-                    self.__tcRunLogger.info("=== Removing DataCenter Successful ===")
+                    self.__logger.info("=== Removing DataCenter Successful ===")
         except Exception as e:
-            self.__tcRunLogger.exception("=== Clean up failed: %s ===" % e)
+            self.__logger.exception("=== Clean up failed: %s ===" % e)
         finally:
             exit(1)
-
 
     def __addToCleanUp(self, type, id):
         if type not in self.__cleanUp.keys():
@@ -78,10 +77,9 @@ class DeployDataCenters(object):
         if type not in self.__cleanUp["order"]:
             self.__cleanUp["order"].append(type)
 
-
     def addHosts(self, hosts, zoneId, podId, clusterId, hypervisor):
         if hosts is None:
-            self.__tcRunLogger.warn("=== Invalid Hosts Information ===")
+            self.__logger.warn("=== Invalid Hosts Information ===")
             return
         failed_cnt = 0
         for host in hosts:
@@ -98,11 +96,11 @@ class DeployDataCenters(object):
                 hostcmd.hypervisor = hypervisor
                 ret = self.__apiClient.addHost(hostcmd)
                 if ret:
-                    self.__tcRunLogger.info("=== Add Host Successful ===")
+                    self.__logger.info("=== Add Host Successful ===")
                     self.__addToCleanUp("Host", ret[0].id)
             except Exception as e:
                 failed_cnt = failed_cnt + 1
-                self.__tcRunLogger.exception("=== Adding Host (%s) Failed: %s ===" % (str(host.url), e))
+                self.__logger.exception("=== Adding Host (%s) Failed: %s ===" % (str(host.url), e))
                 if failed_cnt == len(hosts):
                     self.__cleanAndExit()
 
@@ -127,15 +125,14 @@ class DeployDataCenters(object):
                 clusterresponse = self.__apiClient.addCluster(clustercmd)
                 if clusterresponse[0].id:
                     clusterId = clusterresponse[0].id
-                    self.__tcRunLogger.debug("Cluster Name : %s Id : %s Created Successfully" % (str(cluster.clustername), str(clusterId)))
+                    self.__logger.debug("Cluster Name : %s Id : %s Created Successfully" % (str(cluster.clustername), str(clusterId)))
                     self.__addToCleanUp("Cluster", clusterId)
                 self.addHosts(cluster.hosts, zoneId, podId, clusterId, cluster.hypervisor)
                 self.waitForHost(zoneId, clusterId)
                 self.createPrimaryStorages(cluster.primaryStorages, zoneId, podId, clusterId)
         except Exception as e:
-            self.__tcRunLogger.exception("=== Cluster %s Creation Failed: %s ===" % (str(cluster.clustername), e))
+            self.__logger.exception("=== Cluster %s Creation Failed: %s ===" % (str(cluster.clustername), e))
             self.__cleanAndExit()
-
 
     def waitForHost(self, zoneId, clusterId):
         """
@@ -154,9 +151,8 @@ class DeployDataCenters(object):
                 sleep(timeout)
                 retry = retry - 1
         except Exception as e:
-            self.__tcRunLogger.exception("=== List Hosts Failed: %s ===" % e)
+            self.__logger.exception("=== List Hosts Failed: %s ===" % e)
             self.__cleanAndExit()
-
 
     def createPrimaryStorages(self, primaryStorages, zoneId, podId=None, clusterId=None):
         try:
@@ -181,12 +177,11 @@ class DeployDataCenters(object):
 
                 ret = self.__apiClient.createStoragePool(primarycmd)
                 if ret.id:
-                    self.__tcRunLogger.info("=== Creating Storage Pool Successful ===")
+                    self.__logger.info("=== Creating Storage Pool Successful ===")
                     self.__addToCleanUp("StoragePool", ret.id)
         except Exception as e:
-            self.__tcRunLogger.exception("=== Create Storage Pool Failed: %s ===" % e)
+            self.__logger.exception("=== Create Storage Pool Failed: %s ===" % e)
             self.__cleanAndExit()
-
 
     def createPods(self, pods, zoneId, networkId=None):
         try:
@@ -203,15 +198,14 @@ class DeployDataCenters(object):
                 createpodResponse = self.__apiClient.createPod(createpod)
                 if createpodResponse.id:
                     podId = createpodResponse.id
-                    self.__tcRunLogger.debug("Pod Name : %s Id : %s Created Successfully" % (str(pod.name), str(podId)))
+                    self.__logger.debug("Pod Name : %s Id : %s Created Successfully" % (str(pod.name), str(podId)))
                     self.__addToCleanUp("Pod", podId)
                 if pod.guestIpRanges is not None and networkId is not None:
                     self.createVlanIpRanges("Basic", pod.guestIpRanges, zoneId, podId, networkId)
                 self.createClusters(pod.clusters, zoneId, podId, vmwareDc=pod.vmwaredc)
         except Exception as e:
-            self.__tcRunLogger.exception("=== Pod: %s Creation Failed: %s ===" % (str(pod.name), e))
+            self.__logger.exception("=== Pod: %s Creation Failed: %s ===" % (str(pod.name), e))
             self.__cleanAndExit()
-
 
     def createVlanIpRanges(self, mode, ipranges, zoneId, podId=None, networkId=None, forvirtualnetwork=None):
         try:
@@ -238,12 +232,11 @@ class DeployDataCenters(object):
                     vlanipcmd.forvirtualnetwork = "true"
                 ret = self.__apiClient.createVlanIpRange(vlanipcmd)
                 if ret.id:
-                    self.__tcRunLogger.info("=== Creating Vlan Ip Range Successful ===")
+                    self.__logger.info("=== Creating Vlan Ip Range Successful ===")
                     self.__addToCleanUp("VlanIpRange", ret.id)
         except Exception as e:
-            self.__tcRunLogger.exception("=== Create Vlan Ip Range Failed: %s ===" % e)
+            self.__logger.exception("=== Create Vlan Ip Range Failed: %s ===" % e)
             self.__cleanAndExit()
-
 
     def createSecondaryStorages(self, secondaryStorages, zoneId):
         try:
@@ -265,12 +258,11 @@ class DeployDataCenters(object):
                     secondarycmd.zoneid = zoneId
                 ret = self.__apiClient.addImageStore(secondarycmd)
                 if ret.id:
-                    self.__tcRunLogger.info("=== Add Image Store Successful ===")
+                    self.__logger.info("=== Add Image Store Successful ===")
                     self.__addToCleanUp("ImageStore", ret.id)
         except Exception as e:
-            self.__tcRunLogger.exception("=== Add Image Store Failed: %s ===" % e)
+            self.__logger.exception("=== Add Image Store Failed: %s ===" % e)
             self.__cleanAndExit()
-
 
     def createCacheStorages(self, cacheStorages, zoneId):
         try:
@@ -291,12 +283,11 @@ class DeployDataCenters(object):
                                                 })
                 ret = self.__apiClient.createSecondaryStagingStore(cachecmd)
                 if ret.id:
-                    self.__tcRunLogger.info("=== Creating Secondary StagingStore Successful ===")
+                    self.__logger.info("=== Creating Secondary StagingStore Successful ===")
                     self.__addToCleanUp("SecondaryStagingStore", ret.id)
         except Exception as e:
-            self.__tcRunLogger.exception("=== Creating Secondary Staging Storage Failed: %s ===" % e)
+            self.__logger.exception("=== Creating Secondary Staging Storage Failed: %s ===" % e)
             self.__cleanAndExit()
-
 
     def createNetworks(self, networks, zoneId):
         try:
@@ -319,13 +310,12 @@ class DeployDataCenters(object):
                 networkcmdresponse = self.__apiClient.createNetwork(networkcmd)
                 if networkcmdresponse.id:
                     networkId = networkcmdresponse.id
-                    self.__tcRunLogger.info("=== Creating Network Name : %s Id : %s Successful ===" % (str(network.name), str(networkId)))
+                    self.__logger.info("=== Creating Network Name : %s Id : %s Successful ===" % (str(network.name), str(networkId)))
                     self.__addToCleanUp("Network", networkId)
                     return networkId
         except Exception as e:
-            self.__tcRunLogger.exception("=== Network : %s Creation Failed: %s ===" % (str(network.name), e))
+            self.__logger.exception("=== Network : %s Creation Failed: %s ===" % (str(network.name), e))
             self.__cleanAndExit()
-
 
     def createPhysicalNetwork(self, net, zoneid):
         try:
@@ -335,14 +325,13 @@ class DeployDataCenters(object):
             phynet.isolationmethods = net.isolationmethods
             phynetwrk = self.__apiClient.createPhysicalNetwork(phynet)
             if phynetwrk.id:
-                self.__tcRunLogger.info("=== Creating Physical Network Name : %s Id : %s Successful ===" % (str(phynet.name), str(phynetwrk.id)))
+                self.__logger.info("=== Creating Physical Network Name : %s Id : %s Successful ===" % (str(phynet.name), str(phynetwrk.id)))
                 self.__addToCleanUp("PhysicalNetwork", phynetwrk.id)
             self.addTrafficTypes(phynetwrk.id, net.traffictypes)
             return phynetwrk
         except Exception as e:
-            self.__tcRunLogger.exception("=== Physical Network Creation Failed ===")
+            self.__logger.exception("=== Physical Network Creation Failed ===")
             self.__cleanAndExit()
-
 
     def updatePhysicalNetwork(self, networkid, state="Enabled", vlan=None):
         try:
@@ -354,9 +343,8 @@ class DeployDataCenters(object):
             ret = self.__apiClient.updatePhysicalNetwork(upnet)
             return ret
         except Exception as e:
-            self.__tcRunLogger.exception("=== Update Physical Network Failed ===")
+            self.__logger.exception("=== Update Physical Network Failed ===")
             self.__cleanAndExit()
-
 
     def enableProvider(self, provider_id):
         try:
@@ -366,11 +354,10 @@ class DeployDataCenters(object):
             upnetprov.state = "Enabled"
             ret = self.__apiClient.updateNetworkServiceProvider(upnetprov)
             if ret.id:
-                self.__tcRunLogger.info("=== Update Network Service Provider Successfull ===")
+                self.__logger.info("=== Update Network Service Provider Successfull ===")
         except Exception as e:
-            self.__tcRunLogger.exception("=== Update Network Service Provider Failed ===")
+            self.__logger.exception("=== Update Network Service Provider Failed ===")
             self.__cleanAndExit()
-
 
     def configureProviders(self, phynetwrk, providers):
         """
@@ -414,31 +401,29 @@ class DeployDataCenters(object):
                     result = self.__apiClient.addNetworkServiceProvider(
                         netprov)
                     if result.id:
-                        self.__tcRunLogger.info("=== AddNetworkServiceProvider Successful ===")
+                        self.__logger.info("=== AddNetworkServiceProvider Successful ===")
                         self.__addToCleanUp("NetworkServiceProvider", result.id)
                     if provider.devices is not None:
                         for device in provider.devices:
                             if provider.name == 'NiciraNvp':
-                                cmd =  addNiciraNvpDevice.addNiciraNvpDeviceCmd()
+                                cmd = addNiciraNvpDevice.addNiciraNvpDeviceCmd()
                                 cmd.hostname = device.hostname
                                 cmd.username = device.username
                                 cmd.password = device.password
                                 cmd.transportzoneuuid = device.transportzoneuuid
                                 cmd.physicalnetworkid = phynetwrk.id
                                 ret = self.__apiClient.addNiciraNvpDevice(cmd)
-                                self.__tcRunLogger.info("=== Add NiciraNvp Successful ===")
+                                self.__logger.info("=== Add NiciraNvp Successful ===")
                                 self.__addToCleanUp("NiciraNvp", ret.id)
                             else:
                                 raise InvalidParameterException("Device %s doesn't match any know provider type" % device)
                     self.enableProvider(result.id)
         except Exception as e:
-            self.__tcRunLogger.exception("=== List Network Service Providers Failed: %s ===" % e)
+            self.__logger.exception("=== List Network Service Providers Failed: %s ===" % e)
             self.__cleanAndExit()
-
 
     def addTrafficTypes(self, physical_network_id, traffictypes):
         [self.addTrafficType(physical_network_id, traffic_type) for traffic_type in traffictypes]
-
 
     def addTrafficType(self, physical_network_id, traffictype):
         try:
@@ -455,13 +440,12 @@ class DeployDataCenters(object):
                 if traffictype.simulator is not None else None
             ret = self.__apiClient.addTrafficType(traffic_type)
             if ret.id:
-                self.__tcRunLogger.info("=== Add TrafficType Successful ===")
+                self.__logger.info("=== Add TrafficType Successful ===")
                 self.__addToCleanUp("TrafficType", ret.id)
                 return ret
         except Exception as e:
-            self.__tcRunLogger.exception("=== Add TrafficType Failed: %s ===" % e)
+            self.__logger.exception("=== Add TrafficType Failed: %s ===" % e)
             self.__cleanAndExit()
-
 
     def enableZone(self, zoneid, allocation_state="Enabled"):
         try:
@@ -470,12 +454,11 @@ class DeployDataCenters(object):
             zoneCmd.allocationstate = allocation_state
             ret = self.__apiClient.updateZone(zoneCmd)
             if ret.id:
-                self.__tcRunLogger.info("=== Enable Zone Successful ===")
+                self.__logger.info("=== Enable Zone Successful ===")
                 return ret
         except Exception as e:
-            self.__tcRunLogger.exception("=== Enable Zone Failed: %s ===" % e)
+            self.__logger.exception("=== Enable Zone Failed: %s ===" % e)
             self.__cleanAndExit()
-
 
     def updateZoneDetails(self, zoneid, details):
         try:
@@ -484,31 +467,29 @@ class DeployDataCenters(object):
             zoneCmd.details = details
             ret = self.__apiClient.updateZone(zoneCmd)
             if ret.id:
-                self.__tcRunLogger.info("=== Update Zone Successful ===")
+                self.__logger.info("=== Update Zone Successful ===")
                 return ret
         except Exception as e:
-            self.__tcRunLogger.exception("=== Update Zone  Failed: %s ===" % e)
+            self.__logger.exception("=== Update Zone  Failed: %s ===" % e)
             self.__cleanAndExit()
-
 
     def createZone(self, zone, rec=0):
         try:
             zoneresponse = self.__apiClient.createZone(zone)
             if zoneresponse.id:
-                self.__tcRunLogger.info("=== Create Zone Successful ===")
-                self.__tcRunLogger.debug("Zone Name : %s Id : %s " % (str(zone.name), str(zoneresponse.id)))
+                self.__logger.info("=== Create Zone Successful ===")
+                self.__logger.debug("Zone Name : %s Id : %s " % (str(zone.name), str(zoneresponse.id)))
                 self.__addToCleanUp("Zone", zoneresponse.id)
                 return zoneresponse.id
             else:
-                self.__tcRunLogger.exception("=== Zone : %s Creation Failed ===" % str(zone.name))
+                self.__logger.exception("=== Zone : %s Creation Failed ===" % str(zone.name))
                 if not rec:
                     zone.name = zone.name + "_" + random_gen()
-                    self.__tcRunLogger.info("=== Recreating Zone With New Name : %s ===" % zone.name)
+                    self.__logger.info("=== Recreating Zone With New Name : %s ===" % zone.name)
                     return self.createZone(zone, 1)
         except Exception as e:
-            self.__tcRunLogger.exception("=== Create Zone Failed: %s ===" % e)
+            self.__logger.exception("=== Create Zone Failed: %s ===" % e)
             return FAILED
-
 
     def createZones(self, zones):
         try:
@@ -527,7 +508,7 @@ class DeployDataCenters(object):
                     zonecmd.guestcidraddress = zone.guestcidraddress
                 zoneId = self.createZone(zonecmd)
                 if zoneId == FAILED:
-                    self.__tcRunLogger.error("=== Zone: %s Creation Failed. So Exiting ===" % str(zone.name))
+                    self.__logger.error("=== Zone: %s Creation Failed. So Exiting ===" % str(zone.name))
                     self.__cleanAndExit()
                 for pnet in zone.physical_networks:
                     phynetwrk = self.createPhysicalNetwork(pnet, zoneId)
@@ -538,7 +519,7 @@ class DeployDataCenters(object):
                     if len(filter(lambda x: x.typ == 'Public', zone.physical_networks[0].traffictypes)) > 0:
                         listnetworkoffering.name = "DefaultSharedNetscalerEIPandELBNetworkOffering"
                     else:
-                         listnetworkoffering.name = "DefaultSharedNetworkOfferingWithSGService"
+                        listnetworkoffering.name = "DefaultSharedNetworkOfferingWithSGService"
                     if zone.networkofferingname is not None:
                         listnetworkoffering.name = zone.networkofferingname
                     listnetworkofferingresponse = self.__apiClient.listNetworkOfferings(listnetworkoffering)
@@ -577,7 +558,7 @@ class DeployDataCenters(object):
                     networkcmdresponse = self.__apiClient.createNetwork(networkcmd)
                     if networkcmdresponse.id:
                         self.__addToCleanUp("Network", networkcmdresponse.id)
-                        self.__tcRunLogger.debug("create Network Successful. NetworkId : %s "% str(networkcmdresponse.id))
+                        self.__logger.debug("create Network Successful. NetworkId : %s "% str(networkcmdresponse.id))
                     self.createPods(zone.pods, zoneId, networkcmdresponse.id)
                 '''Note: Swift needs cache storage first'''
                 self.createCacheStorages(zone.cacheStorages, zoneId)
@@ -593,8 +574,7 @@ class DeployDataCenters(object):
                     self.updateZoneDetails(zoneId, det)
             return
         except Exception as e:
-            self.__tcRunLogger.exception("=== Create Zones Failed: %e ===" % e)
-
+            self.__logger.exception("=== Create Zones Failed: %e ===" % e)
 
     def isEipElbZone(self, zone):
         if (zone.networktype == "Basic"
@@ -602,14 +582,12 @@ class DeployDataCenters(object):
             return True
         return False
 
-
     def setClient(self):
         '''
         @Name : setClient
         @Desc : Sets the API Client retrieved from test client
         '''
-        self.__apiClient = self.__testClient.getApiClient()
-
+        self.__apiClient = self.__test_client.getApiClient()
 
     def updateConfiguration(self, globalCfg):
         try:
@@ -621,16 +599,15 @@ class DeployDataCenters(object):
                 updateCfg.value = config.value
                 ret = self.__apiClient.updateConfiguration(updateCfg)
                 if ret.id:
-                    self.__tcRunLogger.info("=== Update Configuration Successfull ===")
+                    self.__logger.info("=== Update Configuration Successfull ===")
         except Exception as e:
-            self.__tcRunLogger.exception("=== Update Configuration Failed: %s ===" % e)
+            self.__logger.exception("=== Update Configuration Failed: %s ===" % e)
             self.__cleanAndExit()
 
     def copyAttributesToCommand(self, source, command):
         map(lambda attr: setattr(command, attr, getattr(source, attr, None)),
             filter(lambda attr: not attr.startswith("__") and attr not in
                    ["required", "isAsync"], dir(command)))
-
 
     def configureS3(self, s3):
         try:
@@ -640,15 +617,15 @@ class DeployDataCenters(object):
             self.copyAttributesToCommand(s3, command)
             ret = self.__apiClient.addS3(command)
             if ret.id:
-                self.__tcRunLogger.info("=== AddS3 Successfull ===")
+                self.__logger.info("=== AddS3 Successfull ===")
                 self.__addToCleanUp("s3", ret.id)
         except Exception as e:
-            self.__tcRunLogger.exception("=== AddS3 Failed: %s ===" % e)
+            self.__logger.exception("=== AddS3 Failed: %s ===" % e)
             self.__cleanAndExit()
 
     def deploy(self):
         try:
-            self.__tcRunLogger.info("=== Deploy DC Started ===")
+            self.__logger.info("=== Deploy DC Started ===")
             '''
             Step1 : Set the Client
             '''
@@ -666,10 +643,10 @@ class DeployDataCenters(object):
             Persist the Configuration to an external file post DC creation
             '''
             self.__persistDcConfig()
-            self.__tcRunLogger.info("=== Deploy DC Successful ===")
+            self.__logger.info("=== Deploy DC Successful ===")
             return SUCCESS
         except Exception as e:
-            self.__tcRunLogger.exception("=== Deploy DC Failed: %s ===" % e)
+            self.__logger.exception("=== Deploy DC Failed: %s ===" % e)
             self.__cleanAndExit()
             return FAILED
 
@@ -678,7 +655,7 @@ class DeleteDataCenters:
 
     '''
     @Desc : Deletes the Data Center using the settings provided.
-            tc_client :Client for deleting the DC.
+            test_client :Client for deleting the DC.
             dc_cfg_file : obj file exported by DeployDataCenter
             when successfully created DC.
                           This file is serialized one containing
@@ -687,13 +664,11 @@ class DeleteDataCenters:
             the dictionary of elements to delete.
             tc_run_logger: Logger to dump log messages.
     '''
-
-    def __init__(self, tc_client, dc_cfg_file=None, dc_cfg=None):
-        self.__dcCfgFile = dc_cfg_file
-        self.__dcCfg = dc_cfg
-        self.__tcRunLogger = MarvinLog('marvin').getLogger()
+    def __init__(self, test_client, cfg):
+        self.__cfg = cfg
+        self.__test_client = test_client
+        self.__logger = MarvinLog('marvin').getLogger()
         self.__apiClient = None
-        self.__testClient = tc_client
 
     def __deleteCmds(self, cmd_name, cmd_obj):
         '''
@@ -737,14 +712,8 @@ class DeleteDataCenters:
                     sleep(30)
         return cmd_obj
 
-
     def __setClient(self):
-        '''
-        @Name : setClient
-        @Desc : Sets the API Client retrieved from test client
-        '''
-        self.__apiClient = self.__testClient.getApiClient()
-
+        self.__apiClient = self.__test_client.getApiClient()
 
     def __cleanEntries(self):
         '''
@@ -753,13 +722,13 @@ class DeleteDataCenters:
         '''
         try:
             ret = FAILED
-            if "order" in self.__dcCfg.keys() and len(self.__dcCfg["order"]):
-                self.__dcCfg["order"].reverse()
-            self.__tcRunLogger.info("=== Clean Up Entries: %s ===" % self.__dcCfg)
-            for type in self.__dcCfg["order"]:
-                self.__tcRunLogger.info("=== CleanUp Started For Type: %s ===" % type)
+            if "order" in self.__cfg.keys() and len(self.__cfg["order"]):
+                self.__cfg["order"].reverse()
+            self.__logger.info("=== Clean Up Entries: %s ===" % self.__cfg)
+            for type in self.__cfg["order"]:
+                self.__logger.info("=== CleanUp Started For Type: %s ===" % type)
                 if type:
-                    temp_ids = self.__dcCfg[type]
+                    temp_ids = self.__cfg[type]
                     ids = [items for items in temp_ids if items]
                     for id in ids:
                         del_mod = "delete" + type
@@ -771,15 +740,14 @@ class DeleteDataCenters:
                         del_func = getattr(self.__apiClient, del_mod)
                         del_cmd_resp = del_func(del_cmd_obj)
                         if del_cmd_resp:
-                            self.__tcRunLogger.error("=== %s CleanUp Failed. ID: %s ===" % (type, id))
+                            self.__logger.error("=== %s CleanUp Failed. ID: %s ===" % (type, id))
                         else:
-                            self.__tcRunLogger.info("=== %s CleanUp Successful. ID : %s ===" % (type, id))
+                            self.__logger.info("=== %s CleanUp Successful. ID : %s ===" % (type, id))
             ret = SUCCESS
         except Exception as e:
-            self.__tcRunLogger.exception("=== Clean Up Entries failed: %s ===" % e)
+            self.__logger.exception("=== Clean Up Entries failed: %s ===" % e)
         finally:
             return ret
-
 
     def removeDataCenter(self):
         '''
@@ -790,95 +758,102 @@ class DeleteDataCenters:
         '''
         try:
             self.__setClient()
-            self.__tcRunLogger.info("=== DeployDC: CleanUp Started ===")
+            self.__logger.info("=== DeployDC: CleanUp Started ===")
             ret = FAILED
-            if self.__dcCfgFile:
-                file_to_read = open(self.__dcCfgFile, 'r')
+            if self.__cfgFile:
+                file_to_read = open(self.__cfgFile, 'r')
                 if file_to_read:
-                    self.__dcCfg = pickle.load(file_to_read)
-            if self.__dcCfg:
+                    self.__cfg = pickle.load(file_to_read)
+            if self.__cfg:
                 ret = self.__cleanEntries()
         except Exception as e:
-            self.__tcRunLogger.exception("=== DeployDC: CleanUp failed: %s ===" % e)
+            self.__logger.exception("=== DeployDC: CleanUp failed: %s ===" % e)
         finally:
             return ret
 
 
+class Application(object):
+
+    def main(self, arguments):
+        options = self.parse_args(arguments)
+        self.validate_options(options)
+        self.run(options)
+
+    def validate_options(self, options):
+        if self.__deploy_and_remove_data_center(options):
+            raise Exception("Can't deploy and remove data center in same run")
+
+        if self.__nothing_to_do(options):
+            raise Exception("No config file (for input nor remove) was specified")
+
+        if self.__options_for_deploy_data_center_are_invalid(options):
+            raise Exception("Invalid input config file path: %s" % options.input)
+
+        if self.__options_for_remove_data_center_are_invalid(options):
+            raise Exception("Invalid remove config file path: %s" % options.remove)
+
+    def parse_args(self, arguments):
+        usage_string = "usage: %prog [options]"
+        parser = OptionParser(usage=usage_string)
+        parser.add_option('-i', '--input',  action='store', default=None, dest='input',  help='marvin config file')
+        parser.add_option('-r', '--remove', action='store', default=None, dest='remove', help='marvin config file')
+
+        (options, _) = parser.parse_args(args=arguments)
+
+        return options
+
+    def create_test_client(self, configuration):
+        test_client = CSTestClient(configuration.mgtSvr[0], configuration.dbSvr)
+        if test_client and test_client.createTestClient() == FAILED:
+            raise Exception("TestClient Creation Failed")
+        return test_client
+
+    def run(self, options):
+        if self.__should_deploy_data_center(options):
+            configuration = configGenerator.getSetupConfig(options.input)
+            test_client = self.create_test_client(configuration)
+            deploy = DeployDataCenters(test_client, configuration)
+            if deploy.deploy() == FAILED:
+                raise Exception("Deploy data center failed")
+        elif self.__should_remove_data_center(options):
+            configuration = configGenerator.getSetupConfig(options.remove)
+            test_client = self.create_test_client(configuration)
+            remove = DeleteDataCenters(test_client, configuration)
+            if remove.removeDataCenter() == FAILED:
+                raise Exception("Removing DataCenter Failed")
+        else:
+            raise Exception("Can't decide on what to do")
+
+    def __should_deploy_data_center(self, options):
+        return self.__option_is_set(options.input) and self.__path_points_to_file(options.input)
+
+    def __options_for_deploy_data_center_are_invalid(self, options):
+        return self.__option_is_set(options.input) and not self.__path_points_to_file(options.input)
+
+    def __should_remove_data_center(self, options):
+        return self.__option_is_set(options.remove) and self.__path_points_to_file(options.remove)
+
+    def __options_for_remove_data_center_are_invalid(self, options):
+        return self.__option_is_set(options.remove) and not self.__path_points_to_file(options.remove)
+
+    def __deploy_and_remove_data_center(self, options):
+        return options.input and options.remove
+
+    def __nothing_to_do(self, options):
+        return options.input is None and options.remove is None
+
+    def __option_is_set(self, flag):
+        return flag is not None
+
+    def __path_points_to_file(self, path):
+        return os.path.isfile(path)
+
+
 if __name__ == "__main__":
-    '''
-    @Desc : This module facilitates the following:
-            1. Deploying DataCenter by using the input provided
-            configuration.
-              EX: python deployDataCenter.py -i <inp-cfg-file>
-            2. Removes a created DataCenter by providing
-            the input configuration file and data center settings file
-              EX: python deployDataCenter.py -i <inp-cfg-file>
-              -r <dc_exported_entries>
-    '''
-    parser = OptionParser()
-    parser.add_option("-i", "--input", action="store",
-                      default=None, dest="input",
-                      help="the path \
-                      where the json config file generated")
-
-    parser.add_option("-r", "--remove", action="store",
-                      default=None, dest="remove",
-                      help="path to file\
-                      where the created dc entries are kept")
-    (options, args) = parser.parse_args()
-
-
-    '''
-    Create the Logger
-    '''
-    from marvin.marvinLog import MarvinLog
-    tc_run_logger = MarvinLog('marvin').getLogger()
-    if (options.input) and not (os.path.isfile(options.input)):
-        tc_run_logger.info("=== Invalid Input Config File Path. Please Check ===")
+    logger = MarvinLog('marvin').getLogger()
+    try:
+        Application().main(sys.argv[1:])
+    except Exception as e:
+        logger.exception("Aborting run due to: %s" % e)
         exit(1)
-
-    '''
-    Verify the input validity
-    '''
-    if options.input is None and options.remove is None:
-        tc_run_logger.error("=== For DeployDataCenter: Please Specify a Valid Input Configuration File ===")
-        tc_run_logger.error("=== For DeleteDataCenters: Please Specify a Valid Input Configuration File and DC Settings ===")
-        exit(1)
-
-    '''
-    Create Test Client
-    '''
-    from marvin.cloudstackTestClient import CSTestClient
-    cfg = configGenerator.getSetupConfig(options.input)
-    obj_tc_client = CSTestClient(cfg.mgtSvr[0], cfg.dbSvr)
-    if obj_tc_client and obj_tc_client.createTestClient() == FAILED:
-        tc_run_logger.error("=== TestClient Creation Failed ===")
-        exit(1)
-
-    '''
-    Verify and continue whether to deploy a DC or remove a DC
-    '''
-
-    if (options.input) and (os.path.isfile(options.input)) and (options.remove is None):
-        '''
-        @Desc : Deploys a Data Center with provided Config
-        '''
-        deploy = DeployDataCenters(obj_tc_client, cfg)
-        if deploy.deploy() == FAILED:
-            tc_run_logger.error("=== Deploy DataCenter Failed ===");
-            exit(1)
-
-    if options.remove and os.path.isfile(options.remove) and options.input:
-        '''
-        @Desc : Removes a Data Center with provided Config
-        '''
-        remove_dc_obj = DeleteDataCenters(obj_tc_client, dc_cfg_file=options.remove)
-        if remove_dc_obj:
-            if remove_dc_obj.removeDataCenter() == FAILED:
-                tc_run_logger.error("=== Removing DataCenter Failed ===")
-                exit(1)
-            else:
-                tc_run_logger.info("=== Removing DataCenter Successful ===")
-
-    # All OK exit with 0 exitcode
     exit(0)
